pyproject.toml
[project]
name = "ll-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "yaoqijun", email = "yaoqijunmail@foxmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
ll-chain = "ll_chain:main"

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
